* fix rd solver for merge
