--- PoseLib/robust/jacobian_impl.h
@@ -1,1631 +1,0 @@
-// Copyright (c) 2021, Viktor Larsson
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
-//
-//     * Redistributions of source code must retain the above copyright
-//       notice, this list of conditions and the following disclaimer.
-//
-//     * Redistributions in binary form must reproduce the above copyright
-//       notice, this list of conditions and the following disclaimer in the
-//       documentation and/or other materials provided with the distribution.
-//
-//     * Neither the name of the copyright holder nor the
-//       names of its contributors may be used to endorse or promote products
-//       derived from this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-// ARE DISCLAIMED. IN NO EVENT SHALL COPYRIGHT HOLDERS OR CONTRIBUTORS BE LIABLE
-// FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-// (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-// LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-// ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-#ifndef POSELIB_JACOBIAN_IMPL_H_
-#define POSELIB_JACOBIAN_IMPL_H_
-
-#include "PoseLib/camera_pose.h"
-#include "PoseLib/misc/camera_models.h"
-#include "PoseLib/misc/essential.h"
-#include "PoseLib/types.h"
-
-namespace poselib {
-
-// For the accumulators we support supplying a vector<double> with point-wise weights for the residuals
-// In case we don't want to have weighted residuals, we can pass UniformWeightVector instead of filling a std::vector
-// with 1.0 The multiplication is then hopefully is optimized away since it always returns 1.0
-class UniformWeightVector {
-  public:
-    UniformWeightVector() {}
-    constexpr double operator[](std::size_t idx) const { return 1.0; }
-};
-class UniformWeightVectors { // this corresponds to std::vector<std::vector<double>> used for generalized cameras etc
-  public:
-    UniformWeightVectors() {}
-    constexpr const UniformWeightVector &operator[](std::size_t idx) const { return w; }
-    const UniformWeightVector w;
-    typedef UniformWeightVector value_type;
-};
-
-template <typename CameraModel, typename LossFunction, typename ResidualWeightVector = UniformWeightVector>
-class CameraJacobianAccumulator {
-  public:
-    CameraJacobianAccumulator(const std::vector<Point2D> &points2D, const std::vector<Point3D> &points3D,
-                              const Camera &cam, const LossFunction &loss,
-                              const ResidualWeightVector &w = ResidualWeightVector())
-        : x(points2D), X(points3D), camera(cam), loss_fn(loss), weights(w) {}
-
-    double residual(const CameraPose &pose) const {
-        double cost = 0;
-        for (size_t i = 0; i < x.size(); ++i) {
-            const Eigen::Vector3d Z = pose.apply(X[i]);
-            // Note this assumes points that are behind the camera will stay behind the camera
-            // during the optimization
-            if (Z(2) < 0)
-                continue;
-            Eigen::Vector2d xp;
-            CameraModel::project(camera.params, Z, &xp);
-            const double r0 = xp(0) - x[i](0);
-            const double r1 = xp(1) - x[i](1);
-            const double r_squared = r0 * r0 + r1 * r1;
-            cost += weights[i] * loss_fn.loss(r_squared);
-        }
-        return cost;
-    }
-
-    // computes J.transpose() * J and J.transpose() * res
-    // Only computes the lower half of JtJ
-    size_t accumulate(const CameraPose &pose, Eigen::Matrix<double, 6, 6> &JtJ,
-                      Eigen::Matrix<double, 6, 1> &Jtr) const {
-        Eigen::Matrix3d R = pose.R();
-        Eigen::Matrix<double, 2, 3> Jproj;
-        size_t num_residuals = 0;
-        for (size_t i = 0; i < x.size(); ++i) {
-            const Eigen::Vector3d Z = R * X[i] + pose.t;
-
-            // Note this assumes points that are behind the camera will stay behind the camera
-            // during the optimization
-            if (Z(2) < 0)
-                continue;
-
-            // Project with intrinsics
-            Eigen::Vector2d zp;
-            camera.project_with_jac(Z, &zp, &Jproj);
-
-            // Setup residual
-            Eigen::Vector2d r = zp - x[i];
-            const double r_squared = r.squaredNorm();
-            const double weight = weights[i] * loss_fn.weight(r_squared);
-
-            if (weight == 0.0) {
-                continue;
-            }
-            num_residuals++;
-
-            // Compute jacobian w.r.t. Z (times R)
-            Eigen::Matrix<double, 2, 3> dZ = Jproj;
-            dZ *= R;
-
-            const double X0 = X[i](0);
-            const double X1 = X[i](1);
-            const double X2 = X[i](2);
-            const double dZtdZ_0_0 = weight * dZ.col(0).dot(dZ.col(0));
-            const double dZtdZ_1_0 = weight * dZ.col(1).dot(dZ.col(0));
-            const double dZtdZ_1_1 = weight * dZ.col(1).dot(dZ.col(1));
-            const double dZtdZ_2_0 = weight * dZ.col(2).dot(dZ.col(0));
-            const double dZtdZ_2_1 = weight * dZ.col(2).dot(dZ.col(1));
-            const double dZtdZ_2_2 = weight * dZ.col(2).dot(dZ.col(2));
-            JtJ(0, 0) += X2 * (X2 * dZtdZ_1_1 - X1 * dZtdZ_2_1) + X1 * (X1 * dZtdZ_2_2 - X2 * dZtdZ_2_1);
-            JtJ(1, 0) += -X2 * (X2 * dZtdZ_1_0 - X0 * dZtdZ_2_1) - X1 * (X0 * dZtdZ_2_2 - X2 * dZtdZ_2_0);
-            JtJ(2, 0) += X1 * (X0 * dZtdZ_2_1 - X1 * dZtdZ_2_0) - X2 * (X0 * dZtdZ_1_1 - X1 * dZtdZ_1_0);
-            JtJ(3, 0) += X1 * dZtdZ_2_0 - X2 * dZtdZ_1_0;
-            JtJ(4, 0) += X1 * dZtdZ_2_1 - X2 * dZtdZ_1_1;
-            JtJ(5, 0) += X1 * dZtdZ_2_2 - X2 * dZtdZ_2_1;
-            JtJ(1, 1) += X2 * (X2 * dZtdZ_0_0 - X0 * dZtdZ_2_0) + X0 * (X0 * dZtdZ_2_2 - X2 * dZtdZ_2_0);
-            JtJ(2, 1) += -X2 * (X1 * dZtdZ_0_0 - X0 * dZtdZ_1_0) - X0 * (X0 * dZtdZ_2_1 - X1 * dZtdZ_2_0);
-            JtJ(3, 1) += X2 * dZtdZ_0_0 - X0 * dZtdZ_2_0;
-            JtJ(4, 1) += X2 * dZtdZ_1_0 - X0 * dZtdZ_2_1;
-            JtJ(5, 1) += X2 * dZtdZ_2_0 - X0 * dZtdZ_2_2;
-            JtJ(2, 2) += X1 * (X1 * dZtdZ_0_0 - X0 * dZtdZ_1_0) + X0 * (X0 * dZtdZ_1_1 - X1 * dZtdZ_1_0);
-            JtJ(3, 2) += X0 * dZtdZ_1_0 - X1 * dZtdZ_0_0;
-            JtJ(4, 2) += X0 * dZtdZ_1_1 - X1 * dZtdZ_1_0;
-            JtJ(5, 2) += X0 * dZtdZ_2_1 - X1 * dZtdZ_2_0;
-            JtJ(3, 3) += dZtdZ_0_0;
-            JtJ(4, 3) += dZtdZ_1_0;
-            JtJ(5, 3) += dZtdZ_2_0;
-            JtJ(4, 4) += dZtdZ_1_1;
-            JtJ(5, 4) += dZtdZ_2_1;
-            JtJ(5, 5) += dZtdZ_2_2;
-            r *= weight;
-            Jtr(0) += (r(0) * (X1 * dZ(0, 2) - X2 * dZ(0, 1)) + r(1) * (X1 * dZ(1, 2) - X2 * dZ(1, 1)));
-            Jtr(1) += (-r(0) * (X0 * dZ(0, 2) - X2 * dZ(0, 0)) - r(1) * (X0 * dZ(1, 2) - X2 * dZ(1, 0)));
-            Jtr(2) += (r(0) * (X0 * dZ(0, 1) - X1 * dZ(0, 0)) + r(1) * (X0 * dZ(1, 1) - X1 * dZ(1, 0)));
-            Jtr(3) += (dZ(0, 0) * r(0) + dZ(1, 0) * r(1));
-            Jtr(4) += (dZ(0, 1) * r(0) + dZ(1, 1) * r(1));
-            Jtr(5) += (dZ(0, 2) * r(0) + dZ(1, 2) * r(1));
-        }
-        return num_residuals;
-    }
-
-    CameraPose step(Eigen::Matrix<double, 6, 1> dp, const CameraPose &pose) const {
-        CameraPose pose_new;
-        // The rotation is parameterized via the lie-rep. and post-multiplication
-        //   i.e. R(delta) = R * expm([delta]_x)
-        pose_new.q = quat_step_post(pose.q, dp.block<3, 1>(0, 0));
-
-        // Translation is parameterized as (negative) shift in position
-        //  i.e. t(delta) = t + R*delta
-        pose_new.t = pose.t + pose.rotate(dp.block<3, 1>(3, 0));
-        return pose_new;
-    }
-    typedef CameraPose param_t;
-    static constexpr size_t num_params = 6;
-
-  private:
-    const std::vector<Point2D> &x;
-    const std::vector<Point3D> &X;
-    const Camera &camera;
-    const LossFunction &loss_fn;
-    const ResidualWeightVector &weights;
-};
-
-template <typename LossFunction, typename ResidualWeightVectors = UniformWeightVectors>
-class GeneralizedCameraJacobianAccumulator {
-  public:
-    GeneralizedCameraJacobianAccumulator(const std::vector<std::vector<Point2D>> &points2D,
-                                         const std::vector<std::vector<Point3D>> &points3D,
-                                         const std::vector<CameraPose> &camera_ext,
-                                         const std::vector<Camera> &camera_int, const LossFunction &l,
-                                         const ResidualWeightVectors &w = ResidualWeightVectors())
-        : num_cams(points2D.size()), x(points2D), X(points3D), rig_poses(camera_ext), cameras(camera_int), loss_fn(l),
-          weights(w) {}
-
-    double residual(const CameraPose &pose) const {
-        double cost = 0.0;
-        for (size_t k = 0; k < num_cams; ++k) {
-            if (x[k].size() == 0) {
-                continue;
-            }
-            const Camera &camera = cameras[k];
-            CameraPose full_pose;
-            full_pose.q = quat_multiply(rig_poses[k].q, pose.q);
-            full_pose.t = rig_poses[k].rotate(pose.t) + rig_poses[k].t;
-
-            switch (camera.model_id) {
-#define SWITCH_CAMERA_MODEL_CASE(Model)                                                                                \
-    case Model::model_id: {                                                                                            \
-        CameraJacobianAccumulator<Model, decltype(loss_fn), typename ResidualWeightVectors::value_type> accum(         \
-            x[k], X[k], cameras[k], loss_fn, weights[k]);                                                              \
-        cost += accum.residual(full_pose);                                                                             \
-        break;                                                                                                         \
-    }
-                SWITCH_CAMERA_MODELS
-
-#undef SWITCH_CAMERA_MODEL_CASE
-            }
-        }
-        return cost;
-    }
-
-    size_t accumulate(const CameraPose &pose, Eigen::Matrix<double, 6, 6> &JtJ,
-                      Eigen::Matrix<double, 6, 1> &Jtr) const {
-        size_t num_residuals = 0;
-
-        for (size_t k = 0; k < num_cams; ++k) {
-            if (x[k].size() == 0) {
-                continue;
-            }
-            const Camera &camera = cameras[k];
-            CameraPose full_pose;
-            full_pose.q = quat_multiply(rig_poses[k].q, pose.q);
-            full_pose.t = rig_poses[k].rotate(pose.t) + rig_poses[k].t;
-
-            switch (camera.model_id) {
-#define SWITCH_CAMERA_MODEL_CASE(Model)                                                                                \
-    case Model::model_id: {                                                                                            \
-        CameraJacobianAccumulator<Model, decltype(loss_fn), typename ResidualWeightVectors::value_type> accum(         \
-            x[k], X[k], cameras[k], loss_fn, weights[k]);                                                              \
-        num_residuals += accum.accumulate(full_pose, JtJ, Jtr);                                                        \
-        break;                                                                                                         \
-    }
-                SWITCH_CAMERA_MODELS
-
-#undef SWITCH_CAMERA_MODEL_CASE
-            }
-        }
-        return num_residuals;
-    }
-
-    CameraPose step(Eigen::Matrix<double, 6, 1> dp, const CameraPose &pose) const {
-        CameraPose pose_new;
-        pose_new.q = quat_step_post(pose.q, dp.block<3, 1>(0, 0));
-        pose_new.t = pose.t + pose.rotate(dp.block<3, 1>(3, 0));
-        return pose_new;
-    }
-    typedef CameraPose param_t;
-    static constexpr size_t num_params = 6;
-
-  private:
-    const size_t num_cams;
-    const std::vector<std::vector<Point2D>> &x;
-    const std::vector<std::vector<Point3D>> &X;
-    const std::vector<CameraPose> &rig_poses;
-    const std::vector<Camera> &cameras;
-    const LossFunction &loss_fn;
-    const ResidualWeightVectors &weights;
-};
-
-template <typename LossFunction, typename ResidualWeightVector = UniformWeightVector> class LineJacobianAccumulator {
-  public:
-    LineJacobianAccumulator(const std::vector<Line2D> &lines2D_, const std::vector<Line3D> &lines3D_,
-                            const LossFunction &loss, const ResidualWeightVector &w = ResidualWeightVector())
-        : lines2D(lines2D_), lines3D(lines3D_), loss_fn(loss), weights(w) {}
-
-    double residual(const CameraPose &pose) const {
-        Eigen::Matrix3d R = pose.R();
-        double cost = 0;
-        for (size_t i = 0; i < lines2D.size(); ++i) {
-            const Eigen::Vector3d Z1 = R * lines3D[i].X1 + pose.t;
-            const Eigen::Vector3d Z2 = R * lines3D[i].X2 + pose.t;
-            Eigen::Vector3d l = Z1.cross(Z2);
-            l /= l.topRows<2>().norm();
-
-            const double r0 = l.dot(lines2D[i].x1.homogeneous());
-            const double r1 = l.dot(lines2D[i].x2.homogeneous());
-            const double r_squared = r0 * r0 + r1 * r1;
-            cost += weights[i] * loss_fn.loss(r_squared);
-        }
-        return cost;
-    }
-
-    // computes J.transpose() * J and J.transpose() * res
-    // Only computes the lower half of JtJ
-    size_t accumulate(const CameraPose &pose, Eigen::Matrix<double, 6, 6> &JtJ,
-                      Eigen::Matrix<double, 6, 1> &Jtr) const {
-
-        Eigen::Matrix3d E, R;
-        R = pose.R();
-        E << pose.t.cross(R.col(0)), pose.t.cross(R.col(1)), pose.t.cross(R.col(2));
-        size_t num_residuals = 0;
-        for (size_t k = 0; k < lines2D.size(); ++k) {
-            const Eigen::Vector3d Z1 = R * lines3D[k].X1 + pose.t;
-            const Eigen::Vector3d Z2 = R * lines3D[k].X2 + pose.t;
-
-            const Eigen::Vector3d X12 = lines3D[k].X1.cross(lines3D[k].X2);
-            const Eigen::Vector3d dX = lines3D[k].X1 - lines3D[k].X2;
-
-            // Projected line
-            const Eigen::Vector3d l = Z1.cross(Z2);
-
-            // Normalized line by first two coordinates
-            Eigen::Vector2d alpha = l.topRows<2>();
-            double beta = l(2);
-            const double n_alpha = alpha.norm();
-            alpha /= n_alpha;
-            beta /= n_alpha;
-
-            // Compute residual
-            Eigen::Vector2d r;
-            r << alpha.dot(lines2D[k].x1) + beta, alpha.dot(lines2D[k].x2) + beta;
-
-            const double r_squared = r.squaredNorm();
-            const double weight = weights[k] * loss_fn.weight(r_squared);
-
-            if (weight == 0.0) {
-                continue;
-            }
-            num_residuals++;
-
-            Eigen::Matrix<double, 3, 6> dl_drt;
-            // Differentiate line with respect to rotation parameters
-            dl_drt.block<1, 3>(0, 0) = E.row(0).cross(dX) - R.row(0).cross(X12);
-            dl_drt.block<1, 3>(1, 0) = E.row(1).cross(dX) - R.row(1).cross(X12);
-            dl_drt.block<1, 3>(2, 0) = E.row(2).cross(dX) - R.row(2).cross(X12);
-            // and translation params
-            dl_drt.block<1, 3>(0, 3) = R.row(0).cross(dX);
-            dl_drt.block<1, 3>(1, 3) = R.row(1).cross(dX);
-            dl_drt.block<1, 3>(2, 3) = R.row(2).cross(dX);
-
-            // Differentiate normalized line w.r.t. original line
-            Eigen::Matrix3d dln_dl;
-            dln_dl.block<2, 2>(0, 0) = (Eigen::Matrix2d::Identity() - alpha * alpha.transpose()) / n_alpha;
-            dln_dl.block<1, 2>(2, 0) = -beta * alpha / n_alpha;
-            dln_dl.block<2, 1>(0, 2).setZero();
-            dln_dl(2, 2) = 1 / n_alpha;
-
-            // Differentiate residual w.r.t. line
-            Eigen::Matrix<double, 2, 3> dr_dl;
-            dr_dl.row(0) << lines2D[k].x1.transpose(), 1.0;
-            dr_dl.row(1) << lines2D[k].x2.transpose(), 1.0;
-
-            Eigen::Matrix<double, 2, 6> J = dr_dl * dln_dl * dl_drt;
-            // Accumulate into JtJ and Jtr
-            Jtr += weight * J.transpose() * r;
-            for (size_t i = 0; i < 6; ++i) {
-                for (size_t j = 0; j <= i; ++j) {
-                    JtJ(i, j) += weight * (J.col(i).dot(J.col(j)));
-                }
-            }
-        }
-        return num_residuals;
-    }
-
-    CameraPose step(Eigen::Matrix<double, 6, 1> dp, const CameraPose &pose) const {
-        CameraPose pose_new;
-        // The rotation is parameterized via the lie-rep. and post-multiplication
-        //   i.e. R(delta) = R * expm([delta]_x)
-        pose_new.q = quat_step_post(pose.q, dp.block<3, 1>(0, 0));
-        // Translation is parameterized as (negative) shift in position
-        //  i.e. t(delta) = t + R*delta
-        pose_new.t = pose.t + pose.rotate(dp.block<3, 1>(3, 0));
-        return pose_new;
-    }
-    typedef CameraPose param_t;
-    static constexpr size_t num_params = 6;
-
-  private:
-    const std::vector<Line2D> &lines2D;
-    const std::vector<Line3D> &lines3D;
-    const LossFunction &loss_fn;
-    const ResidualWeightVector &weights;
-};
-
-template <typename PointLossFunction, typename LineLossFunction, typename PointResidualsVector = UniformWeightVector,
-          typename LineResidualsVector = UniformWeightVector>
-class PointLineJacobianAccumulator {
-  public:
-    PointLineJacobianAccumulator(const std::vector<Point2D> &points2D, const std::vector<Point3D> &points3D,
-                                 const std::vector<Line2D> &lines2D, const std::vector<Line3D> &lines3D,
-                                 const PointLossFunction &l_point, const LineLossFunction &l_line,
-                                 const PointResidualsVector &weights_pts = PointResidualsVector(),
-                                 const LineResidualsVector &weights_l = LineResidualsVector())
-        : pts_accum(points2D, points3D, trivial_camera, l_point, weights_pts),
-          line_accum(lines2D, lines3D, l_line, weights_l) {
-        trivial_camera.model_id = NullCameraModel::model_id;
-    }
-
-    double residual(const CameraPose &pose) const { return pts_accum.residual(pose) + line_accum.residual(pose); }
-
-    size_t accumulate(const CameraPose &pose, Eigen::Matrix<double, 6, 6> &JtJ,
-                      Eigen::Matrix<double, 6, 1> &Jtr) const {
-        return pts_accum.accumulate(pose, JtJ, Jtr) + line_accum.accumulate(pose, JtJ, Jtr);
-    }
-
-    CameraPose step(Eigen::Matrix<double, 6, 1> dp, const CameraPose &pose) const {
-        // Both CameraJacobianAccumulator and LineJacobianAccumulator have the same step!
-        CameraPose pose_new;
-        pose_new.q = quat_step_post(pose.q, dp.block<3, 1>(0, 0));
-        pose_new.t = pose.t + pose.rotate(dp.block<3, 1>(3, 0));
-        return pose_new;
-    }
-    typedef CameraPose param_t;
-    static constexpr size_t num_params = 6;
-
-  private:
-    Camera trivial_camera;
-    CameraJacobianAccumulator<NullCameraModel, PointLossFunction, PointResidualsVector> pts_accum;
-    LineJacobianAccumulator<LineLossFunction, LineResidualsVector> line_accum;
-};
-
-template <typename LossFunction, typename ResidualWeightVector = UniformWeightVector>
-class RelativePoseJacobianAccumulator {
-  public:
-    RelativePoseJacobianAccumulator(const std::vector<Point2D> &points2D_1, const std::vector<Point2D> &points2D_2,
-                                    const LossFunction &l, const ResidualWeightVector &w = ResidualWeightVector())
-        : x1(points2D_1), x2(points2D_2), loss_fn(l), weights(w) {}
-
-    double residual(const CameraPose &pose) const {
-        Eigen::Matrix3d E;
-        essential_from_motion(pose, &E);
-
-        double cost = 0.0;
-        for (size_t k = 0; k < x1.size(); ++k) {
-            double C = x2[k].homogeneous().dot(E * x1[k].homogeneous());
-            double nJc_sq = (E.block<2, 3>(0, 0) * x1[k].homogeneous()).squaredNorm() +
-                            (E.block<3, 2>(0, 0).transpose() * x2[k].homogeneous()).squaredNorm();
-
-            double r2 = (C * C) / nJc_sq;
-            cost += weights[k] * loss_fn.loss(r2);
-        }
-
-        return cost;
-    }
-
-    size_t accumulate(const CameraPose &pose, Eigen::Matrix<double, 5, 5> &JtJ, Eigen::Matrix<double, 5, 1> &Jtr) {
-        // We start by setting up a basis for the updates in the translation (orthogonal to t)
-        // We find the minimum element of t and cross product with the corresponding basis vector.
-        // (this ensures that the first cross product is not close to the zero vector)
-        if (std::abs(pose.t.x()) < std::abs(pose.t.y())) {
-            // x < y
-            if (std::abs(pose.t.x()) < std::abs(pose.t.z())) {
-                tangent_basis.col(0) = pose.t.cross(Eigen::Vector3d::UnitX()).normalized();
-            } else {
-                tangent_basis.col(0) = pose.t.cross(Eigen::Vector3d::UnitZ()).normalized();
-            }
-        } else {
-            // x > y
-            if (std::abs(pose.t.y()) < std::abs(pose.t.z())) {
-                tangent_basis.col(0) = pose.t.cross(Eigen::Vector3d::UnitY()).normalized();
-            } else {
-                tangent_basis.col(0) = pose.t.cross(Eigen::Vector3d::UnitZ()).normalized();
-            }
-        }
-        tangent_basis.col(1) = tangent_basis.col(0).cross(pose.t).normalized();
-
-        Eigen::Matrix3d E, R;
-        R = pose.R();
-        essential_from_motion(pose, &E);
-
-        // Matrices contain the jacobians of E w.r.t. the rotation and translation parameters
-        Eigen::Matrix<double, 9, 3> dR;
-        Eigen::Matrix<double, 9, 2> dt;
-
-        // Each column is vec(E*skew(e_k)) where e_k is k:th basis vector
-        dR.block<3, 1>(0, 0).setZero();
-        dR.block<3, 1>(0, 1) = -E.col(2);
-        dR.block<3, 1>(0, 2) = E.col(1);
-        dR.block<3, 1>(3, 0) = E.col(2);
-        dR.block<3, 1>(3, 1).setZero();
-        dR.block<3, 1>(3, 2) = -E.col(0);
-        dR.block<3, 1>(6, 0) = -E.col(1);
-        dR.block<3, 1>(6, 1) = E.col(0);
-        dR.block<3, 1>(6, 2).setZero();
-
-        // Each column is vec(skew(tangent_basis[k])*R)
-        dt.block<3, 1>(0, 0) = tangent_basis.col(0).cross(R.col(0));
-        dt.block<3, 1>(0, 1) = tangent_basis.col(1).cross(R.col(0));
-        dt.block<3, 1>(3, 0) = tangent_basis.col(0).cross(R.col(1));
-        dt.block<3, 1>(3, 1) = tangent_basis.col(1).cross(R.col(1));
-        dt.block<3, 1>(6, 0) = tangent_basis.col(0).cross(R.col(2));
-        dt.block<3, 1>(6, 1) = tangent_basis.col(1).cross(R.col(2));
-
-        size_t num_residuals = 0;
-        for (size_t k = 0; k < x1.size(); ++k) {
-            double C = x2[k].homogeneous().dot(E * x1[k].homogeneous());
-
-            // J_C is the Jacobian of the epipolar constraint w.r.t. the image points
-            Eigen::Vector4d J_C;
-            J_C << E.block<3, 2>(0, 0).transpose() * x2[k].homogeneous(), E.block<2, 3>(0, 0) * x1[k].homogeneous();
-            const double nJ_C = J_C.norm();
-            const double inv_nJ_C = 1.0 / nJ_C;
-            const double r = C * inv_nJ_C;
-
-            // Compute weight from robust loss function (used in the IRLS)
-            const double weight = weights[k] * loss_fn.weight(r * r);
-            if (weight == 0.0) {
-                continue;
-            }
-            num_residuals++;
-
-            // Compute Jacobian of Sampson error w.r.t the fundamental/essential matrix (3x3)
-            Eigen::Matrix<double, 1, 9> dF;
-            dF << x1[k](0) * x2[k](0), x1[k](0) * x2[k](1), x1[k](0), x1[k](1) * x2[k](0), x1[k](1) * x2[k](1),
-                x1[k](1), x2[k](0), x2[k](1), 1.0;
-            const double s = C * inv_nJ_C * inv_nJ_C;
-            dF(0) -= s * (J_C(2) * x1[k](0) + J_C(0) * x2[k](0));
-            dF(1) -= s * (J_C(3) * x1[k](0) + J_C(0) * x2[k](1));
-            dF(2) -= s * (J_C(0));
-            dF(3) -= s * (J_C(2) * x1[k](1) + J_C(1) * x2[k](0));
-            dF(4) -= s * (J_C(3) * x1[k](1) + J_C(1) * x2[k](1));
-            dF(5) -= s * (J_C(1));
-            dF(6) -= s * (J_C(2));
-            dF(7) -= s * (J_C(3));
-            dF *= inv_nJ_C;
-
-            // and then w.r.t. the pose parameters (rotation + tangent basis for translation)
-            Eigen::Matrix<double, 1, 5> J;
-            J.block<1, 3>(0, 0) = dF * dR;
-            J.block<1, 2>(0, 3) = dF * dt;
-
-            // Accumulate into JtJ and Jtr
-            Jtr += weight * C * inv_nJ_C * J.transpose();
-            JtJ(0, 0) += weight * (J(0) * J(0));
-            JtJ(1, 0) += weight * (J(1) * J(0));
-            JtJ(1, 1) += weight * (J(1) * J(1));
-            JtJ(2, 0) += weight * (J(2) * J(0));
-            JtJ(2, 1) += weight * (J(2) * J(1));
-            JtJ(2, 2) += weight * (J(2) * J(2));
-            JtJ(3, 0) += weight * (J(3) * J(0));
-            JtJ(3, 1) += weight * (J(3) * J(1));
-            JtJ(3, 2) += weight * (J(3) * J(2));
-            JtJ(3, 3) += weight * (J(3) * J(3));
-            JtJ(4, 0) += weight * (J(4) * J(0));
-            JtJ(4, 1) += weight * (J(4) * J(1));
-            JtJ(4, 2) += weight * (J(4) * J(2));
-            JtJ(4, 3) += weight * (J(4) * J(3));
-            JtJ(4, 4) += weight * (J(4) * J(4));
-        }
-        return num_residuals;
-    }
-
-    CameraPose step(Eigen::Matrix<double, 5, 1> dp, const CameraPose &pose) const {
-        CameraPose pose_new;
-        pose_new.q = quat_step_post(pose.q, dp.block<3, 1>(0, 0));
-        pose_new.t = pose.t + tangent_basis * dp.block<2, 1>(3, 0);
-        return pose_new;
-    }
-    typedef CameraPose param_t;
-    static constexpr size_t num_params = 5;
-
-  private:
-    const std::vector<Point2D> &x1;
-    const std::vector<Point2D> &x2;
-    const LossFunction &loss_fn;
-    const ResidualWeightVector &weights;
-    Eigen::Matrix<double, 3, 2> tangent_basis;
-};
-
-template <typename LossFunction, typename ResidualWeightVector = UniformWeightVector>
-class SharedFocalRelativePoseJacobianAccumulator {
-  public:
-    SharedFocalRelativePoseJacobianAccumulator(const std::vector<Point2D> &points2D_1,
-                                               const std::vector<Point2D> &points2D_2, const LossFunction &l,
-                                               const ResidualWeightVector &w = ResidualWeightVector())
-        : x1(points2D_1), x2(points2D_2), loss_fn(l), weights(w) {}
-
-    double residual(const ImagePair &image_pair) const {
-        Eigen::Matrix3d E;
-        essential_from_motion(image_pair.pose, &E);
-        Eigen::Matrix3d K_inv;
-        K_inv << 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, image_pair.camera1.focal();
-
-        Eigen::Matrix3d F = K_inv * (E * K_inv);
-
-        double cost = 0.0;
-        for (size_t k = 0; k < x1.size(); ++k) {
-            double C = x2[k].homogeneous().dot(F * x1[k].homogeneous());
-            double nJc_sq = (F.block<2, 3>(0, 0) * x1[k].homogeneous()).squaredNorm() +
-                            (F.block<3, 2>(0, 0).transpose() * x2[k].homogeneous()).squaredNorm();
-
-            double r2 = (C * C) / nJc_sq;
-            cost += weights[k] * loss_fn.loss(r2);
-        }
-
-        return cost;
-    }
-
-    size_t accumulate(const ImagePair &image_pair, Eigen::Matrix<double, 6, 6> &JtJ, Eigen::Matrix<double, 6, 1> &Jtr) {
-        // We start by setting up a basis for the updates in the translation (orthogonal to t)
-        // We find the minimum element of t and cross product with the corresponding basis vector.
-        // (this ensures that the first cross product is not close to the zero vector)
-        if (std::abs(image_pair.pose.t.x()) < std::abs(image_pair.pose.t.y())) {
-            // x < y
-            if (std::abs(image_pair.pose.t.x()) < std::abs(image_pair.pose.t.z())) {
-                tangent_basis.col(0) = image_pair.pose.t.cross(Eigen::Vector3d::UnitX()).normalized();
-            } else {
-                tangent_basis.col(0) = image_pair.pose.t.cross(Eigen::Vector3d::UnitZ()).normalized();
-            }
-        } else {
-            // x > y
-            if (std::abs(image_pair.pose.t.y()) < std::abs(image_pair.pose.t.z())) {
-                tangent_basis.col(0) = image_pair.pose.t.cross(Eigen::Vector3d::UnitY()).normalized();
-            } else {
-                tangent_basis.col(0) = image_pair.pose.t.cross(Eigen::Vector3d::UnitZ()).normalized();
-            }
-        }
-        tangent_basis.col(1) = tangent_basis.col(0).cross(image_pair.pose.t).normalized();
-
-        double focal = image_pair.camera1.focal();
-        Eigen::Matrix3d K_inv;
-        K_inv << 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, focal;
-
-        Eigen::Matrix3d E, R;
-        R = image_pair.pose.R();
-        essential_from_motion(image_pair.pose, &E);
-        Eigen::Matrix3d F = K_inv * (E * K_inv);
-
-        // Matrices contain the jacobians of E w.r.t. the rotation and translation parameters
-        Eigen::Matrix<double, 9, 3> dR;
-        Eigen::Matrix<double, 9, 2> dt;
-
-        // Each column is vec(E*skew(e_k)) where e_k is k:th basis vector
-        dR.block<3, 1>(0, 0).setZero();
-        dR.block<3, 1>(0, 1) = -E.col(2);
-        dR.block<3, 1>(0, 2) = E.col(1);
-        dR.block<3, 1>(3, 0) = E.col(2);
-        dR.block<3, 1>(3, 1).setZero();
-        dR.block<3, 1>(3, 2) = -E.col(0);
-        dR.block<3, 1>(6, 0) = -E.col(1);
-        dR.block<3, 1>(6, 1) = E.col(0);
-        dR.block<3, 1>(6, 2).setZero();
-
-        dR.row(2) *= focal;
-        dR.row(5) *= focal;
-        dR.row(6) *= focal;
-        dR.row(7) *= focal;
-        dR.row(8) *= focal * focal;
-
-        // Each column is vec(skew(tangent_basis[k])*R)
-        dt.block<3, 1>(0, 0) = tangent_basis.col(0).cross(R.col(0));
-        dt.block<3, 1>(0, 1) = tangent_basis.col(1).cross(R.col(0));
-        dt.block<3, 1>(3, 0) = tangent_basis.col(0).cross(R.col(1));
-        dt.block<3, 1>(3, 1) = tangent_basis.col(1).cross(R.col(1));
-        dt.block<3, 1>(6, 0) = tangent_basis.col(0).cross(R.col(2));
-        dt.block<3, 1>(6, 1) = tangent_basis.col(1).cross(R.col(2));
-
-        dt.row(2) *= focal;
-        dt.row(5) *= focal;
-        dt.row(6) *= focal;
-        dt.row(7) *= focal;
-        dt.row(8) *= focal * focal;
-
-        Eigen::Matrix<double, 9, 1> df;
-
-        df << 0.0, 0.0, E(2, 0), 0.0, 0.0, E(2, 1), E(0, 2), E(1, 2), 2 * E(2, 2) * focal;
-
-        size_t num_residuals = 0;
-
-        for (size_t k = 0; k < x1.size(); ++k) {
-            double C = x2[k].homogeneous().dot(F * x1[k].homogeneous());
-
-            // J_C is the Jacobian of the epipolar constraint w.r.t. the image points
-            Eigen::Vector4d J_C;
-            J_C << F.block<3, 2>(0, 0).transpose() * x2[k].homogeneous(), F.block<2, 3>(0, 0) * x1[k].homogeneous();
-            const double nJ_C = J_C.norm();
-            const double inv_nJ_C = 1.0 / nJ_C;
-            const double r = C * inv_nJ_C;
-
-            // Compute weight from robust loss function (used in the IRLS)
-            const double weight = weights[k] * loss_fn.weight(r * r);
-            if (weight == 0.0) {
-                continue;
-            }
-            num_residuals++;
-
-            // Compute Jacobian of Sampson error w.r.t the fundamental/essential matrix (3x3)
-            Eigen::Matrix<double, 1, 9> dF;
-            dF << x1[k](0) * x2[k](0), x1[k](0) * x2[k](1), x1[k](0), x1[k](1) * x2[k](0), x1[k](1) * x2[k](1),
-                x1[k](1), x2[k](0), x2[k](1), 1.0;
-            const double s = C * inv_nJ_C * inv_nJ_C;
-            dF(0) -= s * (J_C(2) * x1[k](0) + J_C(0) * x2[k](0));
-            dF(1) -= s * (J_C(3) * x1[k](0) + J_C(0) * x2[k](1));
-            dF(2) -= s * (J_C(0));
-            dF(3) -= s * (J_C(2) * x1[k](1) + J_C(1) * x2[k](0));
-            dF(4) -= s * (J_C(3) * x1[k](1) + J_C(1) * x2[k](1));
-            dF(5) -= s * (J_C(1));
-            dF(6) -= s * (J_C(2));
-            dF(7) -= s * (J_C(3));
-            dF *= inv_nJ_C;
-
-            // and then w.r.t. the pose parameters (rotation + tangent basis for translation)
-            Eigen::Matrix<double, 1, 6> J;
-            J.block<1, 3>(0, 0) = dF * dR;
-            J.block<1, 2>(0, 3) = dF * dt;
-            J(0, 5) = dF * df;
-
-            // Accumulate into JtJ and Jtr
-            Jtr += weight * C * inv_nJ_C * J.transpose();
-            JtJ(0, 0) += weight * (J(0) * J(0));
-            JtJ(1, 0) += weight * (J(1) * J(0));
-            JtJ(1, 1) += weight * (J(1) * J(1));
-            JtJ(2, 0) += weight * (J(2) * J(0));
-            JtJ(2, 1) += weight * (J(2) * J(1));
-            JtJ(2, 2) += weight * (J(2) * J(2));
-            JtJ(3, 0) += weight * (J(3) * J(0));
-            JtJ(3, 1) += weight * (J(3) * J(1));
-            JtJ(3, 2) += weight * (J(3) * J(2));
-            JtJ(3, 3) += weight * (J(3) * J(3));
-            JtJ(4, 0) += weight * (J(4) * J(0));
-            JtJ(4, 1) += weight * (J(4) * J(1));
-            JtJ(4, 2) += weight * (J(4) * J(2));
-            JtJ(4, 3) += weight * (J(4) * J(3));
-            JtJ(4, 4) += weight * (J(4) * J(4));
-            JtJ(5, 0) += weight * (J(5) * J(0));
-            JtJ(5, 1) += weight * (J(5) * J(1));
-            JtJ(5, 2) += weight * (J(5) * J(2));
-            JtJ(5, 3) += weight * (J(5) * J(3));
-            JtJ(5, 4) += weight * (J(5) * J(4));
-            JtJ(5, 5) += weight * (J(5) * J(5));
-        }
-        return num_residuals;
-    }
-
-    ImagePair step(Eigen::Matrix<double, 6, 1> dp, const ImagePair &image_pair) const {
-        CameraPose pose_new;
-        pose_new.q = quat_step_post(image_pair.pose.q, dp.block<3, 1>(0, 0));
-        pose_new.t = image_pair.pose.t + tangent_basis * dp.block<2, 1>(3, 0);
-
-        Camera camera_new =
-            Camera("SIMPLE_PINHOLE",
-                   std::vector<double>{std::max(image_pair.camera1.focal() + dp(5, 0), 0.0), 0.0, 0.0}, -1, -1);
-        ImagePair calib_pose_new(pose_new, camera_new, camera_new);
-        return calib_pose_new;
-    }
-    typedef ImagePair param_t;
-    static constexpr size_t num_params = 6;
-
-  private:
-    const std::vector<Point2D> &x1;
-    const std::vector<Point2D> &x2;
-    const LossFunction &loss_fn;
-    const ResidualWeightVector &weights;
-    Eigen::Matrix<double, 3, 2> tangent_basis;
-};
-
-template <typename LossFunction, typename ResidualWeightVectors = UniformWeightVectors>
-class GeneralizedRelativePoseJacobianAccumulator {
-  public:
-    GeneralizedRelativePoseJacobianAccumulator(const std::vector<PairwiseMatches> &pairwise_matches,
-                                               const std::vector<CameraPose> &camera1_ext,
-                                               const std::vector<CameraPose> &camera2_ext, const LossFunction &l,
-                                               const ResidualWeightVectors &w = ResidualWeightVectors())
-        : matches(pairwise_matches), rig1_poses(camera1_ext), rig2_poses(camera2_ext), loss_fn(l), weights(w) {}
-
-    double residual(const CameraPose &pose) const {
-        double cost = 0.0;
-        for (size_t match_k = 0; match_k < matches.size(); ++match_k) {
-            const PairwiseMatches &m = matches[match_k];
-            Eigen::Vector4d q1 = rig1_poses[m.cam_id1].q;
-            Eigen::Vector3d t1 = rig1_poses[m.cam_id1].t;
-
-            Eigen::Vector4d q2 = rig2_poses[m.cam_id2].q;
-            Eigen::Vector3d t2 = rig2_poses[m.cam_id2].t;
-
-            CameraPose relpose;
-            relpose.q = quat_multiply(q2, quat_multiply(pose.q, quat_conj(q1)));
-            relpose.t = t2 + quat_rotate(q2, pose.t) - relpose.rotate(t1);
-            RelativePoseJacobianAccumulator<LossFunction, typename ResidualWeightVectors::value_type> accum(
-                m.x1, m.x2, loss_fn, weights[match_k]);
-            cost += accum.residual(relpose);
-        }
-        return cost;
-    }
-
-    size_t accumulate(const CameraPose &pose, Eigen::Matrix<double, 6, 6> &JtJ,
-                      Eigen::Matrix<double, 6, 1> &Jtr) const {
-        Eigen::Matrix3d R = pose.R();
-        size_t num_residuals = 0;
-        for (size_t match_k = 0; match_k < matches.size(); ++match_k) {
-            const PairwiseMatches &m = matches[match_k];
-
-            // Cameras are
-            // [R1 t1]
-            // [R2 t2] * [R t; 0 1] = [R2*R t2+R2*t]
-
-            // Relative pose is
-            // [R2*R*R1' t2+R2*t-R2*R*R1'*t1]
-            // Essential matrix is
-            // [t2]_x*R2*R*R1' + [R2*t]_x*R2*R*R1' - R2*R*R1'*[t1]_x
-
-            Eigen::Vector4d q1 = rig1_poses[m.cam_id1].q;
-            Eigen::Matrix3d R1 = quat_to_rotmat(q1);
-            Eigen::Vector3d t1 = rig1_poses[m.cam_id1].t;
-
-            Eigen::Vector4d q2 = rig2_poses[m.cam_id2].q;
-            Eigen::Matrix3d R2 = quat_to_rotmat(q2);
-            Eigen::Vector3d t2 = rig2_poses[m.cam_id2].t;
-
-            CameraPose relpose;
-            relpose.q = quat_multiply(q2, quat_multiply(pose.q, quat_conj(q1)));
-            relpose.t = t2 + R2 * pose.t - relpose.rotate(t1);
-            Eigen::Matrix3d E;
-            essential_from_motion(relpose, &E);
-
-            Eigen::Matrix3d R2R = R2 * R;
-            Eigen::Vector3d Rt = R.transpose() * pose.t;
-
-            // The messy expressions below compute
-            // dRdw = [vec(S1) vec(S2) vec(S3)];
-            // dR = (kron(R1,skew(t2)*R2R+ R2*skew(t)*R) + kron(skew(t1)*R1,R2*R)) * dRdw
-            // dt = [vec(R2*R*S1*R1.') vec(R2*R*S2*R1.') vec(R2*R*S3*R1.')]
-
-            // TODO: Replace with something nice
-            Eigen::Matrix<double, 9, 3> dR;
-            Eigen::Matrix<double, 9, 3> dt;
-            dR(0, 0) = R2R(0, 1) * (R1(1, 2) * t1(2) - R1(2, 2) * t1(1)) -
-                       R2R(0, 2) * (R1(1, 1) * t1(2) - R1(2, 1) * t1(1)) +
-                       R1(0, 1) * (R2R(0, 0) * Rt(1) - R2R(0, 1) * Rt(0) - R2R(1, 2) * t2(2) + R2R(2, 2) * t2(1)) +
-                       R1(0, 2) * (R2R(0, 0) * Rt(2) - R2R(0, 2) * Rt(0) + R2R(1, 1) * t2(2) - R2R(2, 1) * t2(1));
-            dR(0, 1) = R2R(0, 2) * (R1(1, 0) * t1(2) - R1(2, 0) * t1(1)) -
-                       R2R(0, 0) * (R1(1, 2) * t1(2) - R1(2, 2) * t1(1)) -
-                       R1(0, 0) * (R2R(0, 0) * Rt(1) - R2R(0, 1) * Rt(0) - R2R(1, 2) * t2(2) + R2R(2, 2) * t2(1)) +
-                       R1(0, 2) * (R2R(0, 1) * Rt(2) - R2R(0, 2) * Rt(1) - R2R(1, 0) * t2(2) + R2R(2, 0) * t2(1));
-            dR(0, 2) = R2R(0, 0) * (R1(1, 1) * t1(2) - R1(2, 1) * t1(1)) -
-                       R2R(0, 1) * (R1(1, 0) * t1(2) - R1(2, 0) * t1(1)) -
-                       R1(0, 0) * (R2R(0, 0) * Rt(2) - R2R(0, 2) * Rt(0) + R2R(1, 1) * t2(2) - R2R(2, 1) * t2(1)) -
-                       R1(0, 1) * (R2R(0, 1) * Rt(2) - R2R(0, 2) * Rt(1) - R2R(1, 0) * t2(2) + R2R(2, 0) * t2(1));
-            dR(1, 0) = R2R(1, 1) * (R1(1, 2) * t1(2) - R1(2, 2) * t1(1)) -
-                       R2R(1, 2) * (R1(1, 1) * t1(2) - R1(2, 1) * t1(1)) +
-                       R1(0, 1) * (R2R(1, 0) * Rt(1) - R2R(1, 1) * Rt(0) + R2R(0, 2) * t2(2) - R2R(2, 2) * t2(0)) +
-                       R1(0, 2) * (R2R(1, 0) * Rt(2) - R2R(1, 2) * Rt(0) - R2R(0, 1) * t2(2) + R2R(2, 1) * t2(0));
-            dR(1, 1) = R2R(1, 2) * (R1(1, 0) * t1(2) - R1(2, 0) * t1(1)) -
-                       R2R(1, 0) * (R1(1, 2) * t1(2) - R1(2, 2) * t1(1)) -
-                       R1(0, 0) * (R2R(1, 0) * Rt(1) - R2R(1, 1) * Rt(0) + R2R(0, 2) * t2(2) - R2R(2, 2) * t2(0)) +
-                       R1(0, 2) * (R2R(1, 1) * Rt(2) - R2R(1, 2) * Rt(1) + R2R(0, 0) * t2(2) - R2R(2, 0) * t2(0));
-            dR(1, 2) = R2R(1, 0) * (R1(1, 1) * t1(2) - R1(2, 1) * t1(1)) -
-                       R2R(1, 1) * (R1(1, 0) * t1(2) - R1(2, 0) * t1(1)) -
-                       R1(0, 0) * (R2R(1, 0) * Rt(2) - R2R(1, 2) * Rt(0) - R2R(0, 1) * t2(2) + R2R(2, 1) * t2(0)) -
-                       R1(0, 1) * (R2R(1, 1) * Rt(2) - R2R(1, 2) * Rt(1) + R2R(0, 0) * t2(2) - R2R(2, 0) * t2(0));
-            dR(2, 0) = R2R(2, 1) * (R1(1, 2) * t1(2) - R1(2, 2) * t1(1)) -
-                       R2R(2, 2) * (R1(1, 1) * t1(2) - R1(2, 1) * t1(1)) +
-                       R1(0, 1) * (R2R(2, 0) * Rt(1) - R2R(2, 1) * Rt(0) - R2R(0, 2) * t2(1) + R2R(1, 2) * t2(0)) +
-                       R1(0, 2) * (R2R(2, 0) * Rt(2) - R2R(2, 2) * Rt(0) + R2R(0, 1) * t2(1) - R2R(1, 1) * t2(0));
-            dR(2, 1) = R2R(2, 2) * (R1(1, 0) * t1(2) - R1(2, 0) * t1(1)) -
-                       R2R(2, 0) * (R1(1, 2) * t1(2) - R1(2, 2) * t1(1)) -
-                       R1(0, 0) * (R2R(2, 0) * Rt(1) - R2R(2, 1) * Rt(0) - R2R(0, 2) * t2(1) + R2R(1, 2) * t2(0)) +
-                       R1(0, 2) * (R2R(2, 1) * Rt(2) - R2R(2, 2) * Rt(1) - R2R(0, 0) * t2(1) + R2R(1, 0) * t2(0));
-            dR(2, 2) = R2R(2, 0) * (R1(1, 1) * t1(2) - R1(2, 1) * t1(1)) -
-                       R2R(2, 1) * (R1(1, 0) * t1(2) - R1(2, 0) * t1(1)) -
-                       R1(0, 0) * (R2R(2, 0) * Rt(2) - R2R(2, 2) * Rt(0) + R2R(0, 1) * t2(1) - R2R(1, 1) * t2(0)) -
-                       R1(0, 1) * (R2R(2, 1) * Rt(2) - R2R(2, 2) * Rt(1) - R2R(0, 0) * t2(1) + R2R(1, 0) * t2(0));
-            dR(3, 0) = R2R(0, 2) * (R1(0, 1) * t1(2) - R1(2, 1) * t1(0)) -
-                       R2R(0, 1) * (R1(0, 2) * t1(2) - R1(2, 2) * t1(0)) +
-                       R1(1, 1) * (R2R(0, 0) * Rt(1) - R2R(0, 1) * Rt(0) - R2R(1, 2) * t2(2) + R2R(2, 2) * t2(1)) +
-                       R1(1, 2) * (R2R(0, 0) * Rt(2) - R2R(0, 2) * Rt(0) + R2R(1, 1) * t2(2) - R2R(2, 1) * t2(1));
-            dR(3, 1) = R2R(0, 0) * (R1(0, 2) * t1(2) - R1(2, 2) * t1(0)) -
-                       R2R(0, 2) * (R1(0, 0) * t1(2) - R1(2, 0) * t1(0)) -
-                       R1(1, 0) * (R2R(0, 0) * Rt(1) - R2R(0, 1) * Rt(0) - R2R(1, 2) * t2(2) + R2R(2, 2) * t2(1)) +
-                       R1(1, 2) * (R2R(0, 1) * Rt(2) - R2R(0, 2) * Rt(1) - R2R(1, 0) * t2(2) + R2R(2, 0) * t2(1));
-            dR(3, 2) = R2R(0, 1) * (R1(0, 0) * t1(2) - R1(2, 0) * t1(0)) -
-                       R2R(0, 0) * (R1(0, 1) * t1(2) - R1(2, 1) * t1(0)) -
-                       R1(1, 0) * (R2R(0, 0) * Rt(2) - R2R(0, 2) * Rt(0) + R2R(1, 1) * t2(2) - R2R(2, 1) * t2(1)) -
-                       R1(1, 1) * (R2R(0, 1) * Rt(2) - R2R(0, 2) * Rt(1) - R2R(1, 0) * t2(2) + R2R(2, 0) * t2(1));
-            dR(4, 0) = R2R(1, 2) * (R1(0, 1) * t1(2) - R1(2, 1) * t1(0)) -
-                       R2R(1, 1) * (R1(0, 2) * t1(2) - R1(2, 2) * t1(0)) +
-                       R1(1, 1) * (R2R(1, 0) * Rt(1) - R2R(1, 1) * Rt(0) + R2R(0, 2) * t2(2) - R2R(2, 2) * t2(0)) +
-                       R1(1, 2) * (R2R(1, 0) * Rt(2) - R2R(1, 2) * Rt(0) - R2R(0, 1) * t2(2) + R2R(2, 1) * t2(0));
-            dR(4, 1) = R2R(1, 0) * (R1(0, 2) * t1(2) - R1(2, 2) * t1(0)) -
-                       R2R(1, 2) * (R1(0, 0) * t1(2) - R1(2, 0) * t1(0)) -
-                       R1(1, 0) * (R2R(1, 0) * Rt(1) - R2R(1, 1) * Rt(0) + R2R(0, 2) * t2(2) - R2R(2, 2) * t2(0)) +
-                       R1(1, 2) * (R2R(1, 1) * Rt(2) - R2R(1, 2) * Rt(1) + R2R(0, 0) * t2(2) - R2R(2, 0) * t2(0));
-            dR(4, 2) = R2R(1, 1) * (R1(0, 0) * t1(2) - R1(2, 0) * t1(0)) -
-                       R2R(1, 0) * (R1(0, 1) * t1(2) - R1(2, 1) * t1(0)) -
-                       R1(1, 0) * (R2R(1, 0) * Rt(2) - R2R(1, 2) * Rt(0) - R2R(0, 1) * t2(2) + R2R(2, 1) * t2(0)) -
-                       R1(1, 1) * (R2R(1, 1) * Rt(2) - R2R(1, 2) * Rt(1) + R2R(0, 0) * t2(2) - R2R(2, 0) * t2(0));
-            dR(5, 0) = R2R(2, 2) * (R1(0, 1) * t1(2) - R1(2, 1) * t1(0)) -
-                       R2R(2, 1) * (R1(0, 2) * t1(2) - R1(2, 2) * t1(0)) +
-                       R1(1, 1) * (R2R(2, 0) * Rt(1) - R2R(2, 1) * Rt(0) - R2R(0, 2) * t2(1) + R2R(1, 2) * t2(0)) +
-                       R1(1, 2) * (R2R(2, 0) * Rt(2) - R2R(2, 2) * Rt(0) + R2R(0, 1) * t2(1) - R2R(1, 1) * t2(0));
-            dR(5, 1) = R2R(2, 0) * (R1(0, 2) * t1(2) - R1(2, 2) * t1(0)) -
-                       R2R(2, 2) * (R1(0, 0) * t1(2) - R1(2, 0) * t1(0)) -
-                       R1(1, 0) * (R2R(2, 0) * Rt(1) - R2R(2, 1) * Rt(0) - R2R(0, 2) * t2(1) + R2R(1, 2) * t2(0)) +
-                       R1(1, 2) * (R2R(2, 1) * Rt(2) - R2R(2, 2) * Rt(1) - R2R(0, 0) * t2(1) + R2R(1, 0) * t2(0));
-            dR(5, 2) = R2R(2, 1) * (R1(0, 0) * t1(2) - R1(2, 0) * t1(0)) -
-                       R2R(2, 0) * (R1(0, 1) * t1(2) - R1(2, 1) * t1(0)) -
-                       R1(1, 0) * (R2R(2, 0) * Rt(2) - R2R(2, 2) * Rt(0) + R2R(0, 1) * t2(1) - R2R(1, 1) * t2(0)) -
-                       R1(1, 1) * (R2R(2, 1) * Rt(2) - R2R(2, 2) * Rt(1) - R2R(0, 0) * t2(1) + R2R(1, 0) * t2(0));
-            dR(6, 0) = R2R(0, 1) * (R1(0, 2) * t1(1) - R1(1, 2) * t1(0)) -
-                       R2R(0, 2) * (R1(0, 1) * t1(1) - R1(1, 1) * t1(0)) +
-                       R1(2, 1) * (R2R(0, 0) * Rt(1) - R2R(0, 1) * Rt(0) - R2R(1, 2) * t2(2) + R2R(2, 2) * t2(1)) +
-                       R1(2, 2) * (R2R(0, 0) * Rt(2) - R2R(0, 2) * Rt(0) + R2R(1, 1) * t2(2) - R2R(2, 1) * t2(1));
-            dR(6, 1) = R2R(0, 2) * (R1(0, 0) * t1(1) - R1(1, 0) * t1(0)) -
-                       R2R(0, 0) * (R1(0, 2) * t1(1) - R1(1, 2) * t1(0)) -
-                       R1(2, 0) * (R2R(0, 0) * Rt(1) - R2R(0, 1) * Rt(0) - R2R(1, 2) * t2(2) + R2R(2, 2) * t2(1)) +
-                       R1(2, 2) * (R2R(0, 1) * Rt(2) - R2R(0, 2) * Rt(1) - R2R(1, 0) * t2(2) + R2R(2, 0) * t2(1));
-            dR(6, 2) = R2R(0, 0) * (R1(0, 1) * t1(1) - R1(1, 1) * t1(0)) -
-                       R2R(0, 1) * (R1(0, 0) * t1(1) - R1(1, 0) * t1(0)) -
-                       R1(2, 0) * (R2R(0, 0) * Rt(2) - R2R(0, 2) * Rt(0) + R2R(1, 1) * t2(2) - R2R(2, 1) * t2(1)) -
-                       R1(2, 1) * (R2R(0, 1) * Rt(2) - R2R(0, 2) * Rt(1) - R2R(1, 0) * t2(2) + R2R(2, 0) * t2(1));
-            dR(7, 0) = R2R(1, 1) * (R1(0, 2) * t1(1) - R1(1, 2) * t1(0)) -
-                       R2R(1, 2) * (R1(0, 1) * t1(1) - R1(1, 1) * t1(0)) +
-                       R1(2, 1) * (R2R(1, 0) * Rt(1) - R2R(1, 1) * Rt(0) + R2R(0, 2) * t2(2) - R2R(2, 2) * t2(0)) +
-                       R1(2, 2) * (R2R(1, 0) * Rt(2) - R2R(1, 2) * Rt(0) - R2R(0, 1) * t2(2) + R2R(2, 1) * t2(0));
-            dR(7, 1) = R2R(1, 2) * (R1(0, 0) * t1(1) - R1(1, 0) * t1(0)) -
-                       R2R(1, 0) * (R1(0, 2) * t1(1) - R1(1, 2) * t1(0)) -
-                       R1(2, 0) * (R2R(1, 0) * Rt(1) - R2R(1, 1) * Rt(0) + R2R(0, 2) * t2(2) - R2R(2, 2) * t2(0)) +
-                       R1(2, 2) * (R2R(1, 1) * Rt(2) - R2R(1, 2) * Rt(1) + R2R(0, 0) * t2(2) - R2R(2, 0) * t2(0));
-            dR(7, 2) = R2R(1, 0) * (R1(0, 1) * t1(1) - R1(1, 1) * t1(0)) -
-                       R2R(1, 1) * (R1(0, 0) * t1(1) - R1(1, 0) * t1(0)) -
-                       R1(2, 0) * (R2R(1, 0) * Rt(2) - R2R(1, 2) * Rt(0) - R2R(0, 1) * t2(2) + R2R(2, 1) * t2(0)) -
-                       R1(2, 1) * (R2R(1, 1) * Rt(2) - R2R(1, 2) * Rt(1) + R2R(0, 0) * t2(2) - R2R(2, 0) * t2(0));
-            dR(8, 0) = R2R(2, 1) * (R1(0, 2) * t1(1) - R1(1, 2) * t1(0)) -
-                       R2R(2, 2) * (R1(0, 1) * t1(1) - R1(1, 1) * t1(0)) +
-                       R1(2, 1) * (R2R(2, 0) * Rt(1) - R2R(2, 1) * Rt(0) - R2R(0, 2) * t2(1) + R2R(1, 2) * t2(0)) +
-                       R1(2, 2) * (R2R(2, 0) * Rt(2) - R2R(2, 2) * Rt(0) + R2R(0, 1) * t2(1) - R2R(1, 1) * t2(0));
-            dR(8, 1) = R2R(2, 2) * (R1(0, 0) * t1(1) - R1(1, 0) * t1(0)) -
-                       R2R(2, 0) * (R1(0, 2) * t1(1) - R1(1, 2) * t1(0)) -
-                       R1(2, 0) * (R2R(2, 0) * Rt(1) - R2R(2, 1) * Rt(0) - R2R(0, 2) * t2(1) + R2R(1, 2) * t2(0)) +
-                       R1(2, 2) * (R2R(2, 1) * Rt(2) - R2R(2, 2) * Rt(1) - R2R(0, 0) * t2(1) + R2R(1, 0) * t2(0));
-            dR(8, 2) = R2R(2, 0) * (R1(0, 1) * t1(1) - R1(1, 1) * t1(0)) -
-                       R2R(2, 1) * (R1(0, 0) * t1(1) - R1(1, 0) * t1(0)) -
-                       R1(2, 0) * (R2R(2, 0) * Rt(2) - R2R(2, 2) * Rt(0) + R2R(0, 1) * t2(1) - R2R(1, 1) * t2(0)) -
-                       R1(2, 1) * (R2R(2, 1) * Rt(2) - R2R(2, 2) * Rt(1) - R2R(0, 0) * t2(1) + R2R(1, 0) * t2(0));
-            dt(0, 0) = R2R(0, 2) * R1(0, 1) - R2R(0, 1) * R1(0, 2);
-            dt(0, 1) = R2R(0, 0) * R1(0, 2) - R2R(0, 2) * R1(0, 0);
-            dt(0, 2) = R2R(0, 1) * R1(0, 0) - R2R(0, 0) * R1(0, 1);
-            dt(1, 0) = R2R(1, 2) * R1(0, 1) - R2R(1, 1) * R1(0, 2);
-            dt(1, 1) = R2R(1, 0) * R1(0, 2) - R2R(1, 2) * R1(0, 0);
-            dt(1, 2) = R2R(1, 1) * R1(0, 0) - R2R(1, 0) * R1(0, 1);
-            dt(2, 0) = R2R(2, 2) * R1(0, 1) - R2R(2, 1) * R1(0, 2);
-            dt(2, 1) = R2R(2, 0) * R1(0, 2) - R2R(2, 2) * R1(0, 0);
-            dt(2, 2) = R2R(2, 1) * R1(0, 0) - R2R(2, 0) * R1(0, 1);
-            dt(3, 0) = R2R(0, 2) * R1(1, 1) - R2R(0, 1) * R1(1, 2);
-            dt(3, 1) = R2R(0, 0) * R1(1, 2) - R2R(0, 2) * R1(1, 0);
-            dt(3, 2) = R2R(0, 1) * R1(1, 0) - R2R(0, 0) * R1(1, 1);
-            dt(4, 0) = R2R(1, 2) * R1(1, 1) - R2R(1, 1) * R1(1, 2);
-            dt(4, 1) = R2R(1, 0) * R1(1, 2) - R2R(1, 2) * R1(1, 0);
-            dt(4, 2) = R2R(1, 1) * R1(1, 0) - R2R(1, 0) * R1(1, 1);
-            dt(5, 0) = R2R(2, 2) * R1(1, 1) - R2R(2, 1) * R1(1, 2);
-            dt(5, 1) = R2R(2, 0) * R1(1, 2) - R2R(2, 2) * R1(1, 0);
-            dt(5, 2) = R2R(2, 1) * R1(1, 0) - R2R(2, 0) * R1(1, 1);
-            dt(6, 0) = R2R(0, 2) * R1(2, 1) - R2R(0, 1) * R1(2, 2);
-            dt(6, 1) = R2R(0, 0) * R1(2, 2) - R2R(0, 2) * R1(2, 0);
-            dt(6, 2) = R2R(0, 1) * R1(2, 0) - R2R(0, 0) * R1(2, 1);
-            dt(7, 0) = R2R(1, 2) * R1(2, 1) - R2R(1, 1) * R1(2, 2);
-            dt(7, 1) = R2R(1, 0) * R1(2, 2) - R2R(1, 2) * R1(2, 0);
-            dt(7, 2) = R2R(1, 1) * R1(2, 0) - R2R(1, 0) * R1(2, 1);
-            dt(8, 0) = R2R(2, 2) * R1(2, 1) - R2R(2, 1) * R1(2, 2);
-            dt(8, 1) = R2R(2, 0) * R1(2, 2) - R2R(2, 2) * R1(2, 0);
-            dt(8, 2) = R2R(2, 1) * R1(2, 0) - R2R(2, 0) * R1(2, 1);
-
-            for (size_t k = 0; k < m.x1.size(); ++k) {
-                double C = m.x2[k].homogeneous().dot(E * m.x1[k].homogeneous());
-
-                // J_C is the Jacobian of the epipolar constraint w.r.t. the image points
-                Eigen::Vector4d J_C;
-                J_C << E.block<3, 2>(0, 0).transpose() * m.x2[k].homogeneous(),
-                    E.block<2, 3>(0, 0) * m.x1[k].homogeneous();
-                const double nJ_C = J_C.norm();
-                const double inv_nJ_C = 1.0 / nJ_C;
-                const double r = C * inv_nJ_C;
-
-                // Compute weight from robust loss function (used in the IRLS)
-                const double weight = weights[match_k][k] * loss_fn.weight(r * r);
-                if (weight == 0.0) {
-                    continue;
-                }
-                num_residuals++;
-
-                // Compute Jacobian of Sampson error w.r.t the fundamental/essential matrix (3x3)
-                Eigen::Matrix<double, 1, 9> dF;
-                dF << m.x1[k](0) * m.x2[k](0), m.x1[k](0) * m.x2[k](1), m.x1[k](0), m.x1[k](1) * m.x2[k](0),
-                    m.x1[k](1) * m.x2[k](1), m.x1[k](1), m.x2[k](0), m.x2[k](1), 1.0;
-                const double s = C * inv_nJ_C * inv_nJ_C;
-                dF(0) -= s * (J_C(2) * m.x1[k](0) + J_C(0) * m.x2[k](0));
-                dF(1) -= s * (J_C(3) * m.x1[k](0) + J_C(0) * m.x2[k](1));
-                dF(2) -= s * (J_C(0));
-                dF(3) -= s * (J_C(2) * m.x1[k](1) + J_C(1) * m.x2[k](0));
-                dF(4) -= s * (J_C(3) * m.x1[k](1) + J_C(1) * m.x2[k](1));
-                dF(5) -= s * (J_C(1));
-                dF(6) -= s * (J_C(2));
-                dF(7) -= s * (J_C(3));
-                dF *= inv_nJ_C;
-
-                // and then w.r.t. the pose parameters
-                Eigen::Matrix<double, 1, 6> J;
-                J.block<1, 3>(0, 0) = dF * dR;
-                J.block<1, 3>(0, 3) = dF * dt;
-
-                // Accumulate into JtJ and Jtr
-                Jtr += weight * C * inv_nJ_C * J.transpose();
-                for (size_t i = 0; i < 6; ++i) {
-                    for (size_t j = 0; j <= i; ++j) {
-                        JtJ(i, j) += weight * (J(i) * J(j));
-                    }
-                }
-            }
-        }
-        return num_residuals;
-    }
-
-    CameraPose step(Eigen::Matrix<double, 6, 1> dp, const CameraPose &pose) const {
-        CameraPose pose_new;
-        pose_new.q = quat_step_post(pose.q, dp.block<3, 1>(0, 0));
-        pose_new.t = pose.t + pose.rotate(dp.block<3, 1>(3, 0));
-        return pose_new;
-    }
-    typedef CameraPose param_t;
-    static constexpr size_t num_params = 6;
-
-  private:
-    const std::vector<PairwiseMatches> &matches;
-    const std::vector<CameraPose> &rig1_poses;
-    const std::vector<CameraPose> &rig2_poses;
-    const LossFunction &loss_fn;
-    const ResidualWeightVectors &weights;
-};
-
-template <typename LossFunction, typename AbsResidualsVector = UniformWeightVector,
-          typename RelResidualsVectors = UniformWeightVectors>
-class HybridPoseJacobianAccumulator {
-  public:
-    HybridPoseJacobianAccumulator(const std::vector<Point2D> &points2D, const std::vector<Point3D> &points3D,
-                                  const std::vector<PairwiseMatches> &pairwise_matches,
-                                  const std::vector<CameraPose> &map_ext, const LossFunction &l,
-                                  const LossFunction &l_epi,
-                                  const AbsResidualsVector &weights_abs = AbsResidualsVector(),
-                                  const RelResidualsVectors &weights_rel = RelResidualsVectors())
-        : abs_pose_accum(points2D, points3D, trivial_camera, l, weights_abs),
-          gen_rel_accum(pairwise_matches, map_ext, trivial_rig, l_epi, weights_rel) {
-        trivial_camera.model_id = NullCameraModel::model_id;
-        trivial_rig.emplace_back();
-    }
-
-    double residual(const CameraPose &pose) const {
-        return abs_pose_accum.residual(pose) + gen_rel_accum.residual(pose);
-    }
-
-    size_t accumulate(const CameraPose &pose, Eigen::Matrix<double, 6, 6> &JtJ,
-                      Eigen::Matrix<double, 6, 1> &Jtr) const {
-        return abs_pose_accum.accumulate(pose, JtJ, Jtr) + gen_rel_accum.accumulate(pose, JtJ, Jtr);
-    }
-
-    CameraPose step(Eigen::Matrix<double, 6, 1> dp, const CameraPose &pose) const {
-        CameraPose pose_new;
-        pose_new.q = quat_step_post(pose.q, dp.block<3, 1>(0, 0));
-        pose_new.t = pose.t + pose.rotate(dp.block<3, 1>(3, 0));
-        return pose_new;
-    }
-    typedef CameraPose param_t;
-    static constexpr size_t num_params = 6;
-
-  private:
-    Camera trivial_camera;
-    std::vector<CameraPose> trivial_rig;
-    CameraJacobianAccumulator<NullCameraModel, LossFunction, AbsResidualsVector> abs_pose_accum;
-    GeneralizedRelativePoseJacobianAccumulator<LossFunction, RelResidualsVectors> gen_rel_accum;
-};
-
-// This is the SVD factorization proposed by Bartoli and Sturm in
-// Non-Linear Estimation of the Fundamental Matrix With Minimal Parameters, PAMI 2004
-// Though we do different updates (lie vs the euler angles used in the original paper)
-struct FactorizedFundamentalMatrix {
-    FactorizedFundamentalMatrix() {}
-    FactorizedFundamentalMatrix(const Eigen::Matrix3d &F) {
-        Eigen::JacobiSVD<Eigen::Matrix3d> svd(F, Eigen::ComputeFullV | Eigen::ComputeFullU);
-        Eigen::Matrix3d U = svd.matrixU();
-        Eigen::Matrix3d V = svd.matrixV();
-        if (U.determinant() < 0) {
-            U = -U;
-        }
-        if (V.determinant() < 0) {
-            V = -V;
-        }
-        qU = rotmat_to_quat(U);
-        qV = rotmat_to_quat(V);
-        Eigen::Vector3d s = svd.singularValues();
-        sigma = s(1) / s(0);
-    }
-    Eigen::Matrix3d F() const {
-        Eigen::Matrix3d U = quat_to_rotmat(qU);
-        Eigen::Matrix3d V = quat_to_rotmat(qV);
-        return U.col(0) * V.col(0).transpose() + sigma * U.col(1) * V.col(1).transpose();
-    }
-
-    Eigen::Vector4d qU, qV;
-    double sigma;
-};
-
-template <typename LossFunction, typename ResidualWeightVector = UniformWeightVector>
-class FundamentalJacobianAccumulator {
-  public:
-    FundamentalJacobianAccumulator(const std::vector<Point2D> &points2D_1, const std::vector<Point2D> &points2D_2,
-                                   const LossFunction &l, const ResidualWeightVector &w = ResidualWeightVector())
-        : x1(points2D_1), x2(points2D_2), loss_fn(l), weights(w) {}
-
-    double residual(const FactorizedFundamentalMatrix &FF) const {
-        Eigen::Matrix3d F = FF.F();
-
-        double cost = 0.0;
-        for (size_t k = 0; k < x1.size(); ++k) {
-            double C = x2[k].homogeneous().dot(F * x1[k].homogeneous());
-            double nJc_sq = (F.block<2, 3>(0, 0) * x1[k].homogeneous()).squaredNorm() +
-                            (F.block<3, 2>(0, 0).transpose() * x2[k].homogeneous()).squaredNorm();
-
-            double r2 = (C * C) / nJc_sq;
-            cost += weights[k] * loss_fn.loss(r2);
-        }
-
-        return cost;
-    }
-
-    double residual(const FactorizedFundamentalMatrix &FF, size_t k) const {
-        Eigen::Matrix3d F = FF.F();
-
-        double cost = 0.0;
-        double C = x2[k].homogeneous().dot(F * x1[k].homogeneous());
-        double nJc_sq = (F.block<2, 3>(0, 0) * x1[k].homogeneous()).squaredNorm() +
-                        (F.block<3, 2>(0, 0).transpose() * x2[k].homogeneous()).squaredNorm();
-
-        double r2 = (C * C) / nJc_sq;
-        cost += weights[k] * loss_fn.loss(r2);
-
-        return cost;
-    }
-
-    size_t accumulate(const FactorizedFundamentalMatrix &FF, Eigen::Matrix<double, 7, 7> &JtJ,
-                      Eigen::Matrix<double, 7, 1> &Jtr) const {
-
-        const Eigen::Matrix3d F = FF.F();
-
-        // Matrices contain the jacobians of F w.r.t. the factorized fundamental matrix (U,V,sigma)
-        const Eigen::Matrix3d U = quat_to_rotmat(FF.qU);
-        const Eigen::Matrix3d V = quat_to_rotmat(FF.qV);
-
-        const Eigen::Matrix3d d_sigma = U.col(1) * V.col(1).transpose();
-        Eigen::Matrix<double, 9, 7> dF_dparams;
-        dF_dparams << 0, F(2, 0), -F(1, 0), 0, F(0, 2), -F(0, 1), d_sigma(0, 0), -F(2, 0), 0, F(0, 0), 0, F(1, 2),
-            -F(1, 1), d_sigma(1, 0), F(1, 0), -F(0, 0), 0, 0, F(2, 2), -F(2, 1), d_sigma(2, 0), 0, F(2, 1), -F(1, 1),
-            -F(0, 2), 0, F(0, 0), d_sigma(0, 1), -F(2, 1), 0, F(0, 1), -F(1, 2), 0, F(1, 0), d_sigma(1, 1), F(1, 1),
-            -F(0, 1), 0, -F(2, 2), 0, F(2, 0), d_sigma(2, 1), 0, F(2, 2), -F(1, 2), F(0, 1), -F(0, 0), 0, d_sigma(0, 2),
-            -F(2, 2), 0, F(0, 2), F(1, 1), -F(1, 0), 0, d_sigma(1, 2), F(1, 2), -F(0, 2), 0, F(2, 1), -F(2, 0), 0,
-            d_sigma(2, 2);
-
-        size_t num_residuals = 0;
-        for (size_t k = 0; k < x1.size(); ++k) {
-            const double C = x2[k].homogeneous().dot(F * x1[k].homogeneous());
-
-            // J_C is the Jacobian of the epipolar constraint w.r.t. the image points
-            Eigen::Vector4d J_C;
-            J_C << F.block<3, 2>(0, 0).transpose() * x2[k].homogeneous(), F.block<2, 3>(0, 0) * x1[k].homogeneous();
-            const double nJ_C = J_C.norm();
-            const double inv_nJ_C = 1.0 / nJ_C;
-            const double r = C * inv_nJ_C;
-
-            // Compute weight from robust loss function (used in the IRLS)
-            const double weight = weights[k] * loss_fn.weight(r * r);
-            if (weight == 0.0) {
-                continue;
-            }
-            num_residuals++;
-
-            // Compute Jacobian of Sampson error w.r.t the fundamental/essential matrix (3x3)
-            Eigen::Matrix<double, 1, 9> dF;
-            dF << x1[k](0) * x2[k](0), x1[k](0) * x2[k](1), x1[k](0), x1[k](1) * x2[k](0), x1[k](1) * x2[k](1),
-                x1[k](1), x2[k](0), x2[k](1), 1.0;
-            const double s = C * inv_nJ_C * inv_nJ_C;
-            dF(0) -= s * (J_C(2) * x1[k](0) + J_C(0) * x2[k](0));
-            dF(1) -= s * (J_C(3) * x1[k](0) + J_C(0) * x2[k](1));
-            dF(2) -= s * (J_C(0));
-            dF(3) -= s * (J_C(2) * x1[k](1) + J_C(1) * x2[k](0));
-            dF(4) -= s * (J_C(3) * x1[k](1) + J_C(1) * x2[k](1));
-            dF(5) -= s * (J_C(1));
-            dF(6) -= s * (J_C(2));
-            dF(7) -= s * (J_C(3));
-            dF *= inv_nJ_C;
-
-            // and then w.r.t. the pose parameters (rotation + tangent basis for translation)
-            Eigen::Matrix<double, 1, 7> J = dF * dF_dparams;
-
-            // Accumulate into JtJ and Jtr
-            Jtr += weight * C * inv_nJ_C * J.transpose();
-            for (size_t i = 0; i < 7; ++i) {
-                for (size_t j = 0; j <= i; ++j) {
-                    JtJ(i, j) += weight * (J(i) * J(j));
-                }
-            }
-        }
-        return num_residuals;
-    }
-
-    FactorizedFundamentalMatrix step(Eigen::Matrix<double, 7, 1> dp, const FactorizedFundamentalMatrix &F) const {
-        FactorizedFundamentalMatrix F_new;
-        F_new.qU = quat_step_pre(F.qU, dp.block<3, 1>(0, 0));
-        F_new.qV = quat_step_pre(F.qV, dp.block<3, 1>(3, 0));
-        F_new.sigma = F.sigma + dp(6);
-        return F_new;
-    }
-    typedef FactorizedFundamentalMatrix param_t;
-    static constexpr size_t num_params = 7;
-
-  private:
-    const std::vector<Point2D> &x1;
-    const std::vector<Point2D> &x2;
-    const LossFunction &loss_fn;
-    const ResidualWeightVector &weights;
-};
-
-template <typename LossFunction, typename ResidualWeightVector = UniformWeightVector>
-class TSampRDFundamentalJacobianAccumulator {
-  public:
-    TSampRDFundamentalJacobianAccumulator(const std::vector<Point2D> &points2D_1,
-                                          const std::vector<Point2D> &points2D_2, const LossFunction &l,
-                                          const ResidualWeightVector &w = ResidualWeightVector())
-        : x1(points2D_1), x2(points2D_2), loss_fn(l), weights(w) {}
-
-    double residual(const ProjectiveImagePair F_cam_pair) const {
-        Eigen::Matrix3d F = F_cam_pair.F;
-
-        double cost = 0.0;
-        for (size_t i = 0; i < x1.size(); ++i) {
-            Eigen::Matrix<double, 3, 1> xu1, xu2;
-            Eigen::Matrix<double, 3, 2> J1, J2;
-            F_cam_pair.camera1.unproject_with_jac(x1[i], &xu1, &J1);
-            F_cam_pair.camera2.unproject_with_jac(x2[i], &xu2, &J2);
-
-            double num = xu2.transpose() * (F * xu1);
-            num *= num;
-
-            double den =
-                (xu2.transpose() * F * J1).squaredNorm() + (xu1.transpose() * F.transpose() * J2).squaredNorm();
-            double r2 = num / den;
-            cost += weights[i] * loss_fn.loss(r2);
-        }
-
-        return cost;
-    }
-
-    size_t accumulate(const ProjectiveImagePair F_cam_pair, Eigen::Matrix<double, 9, 9> &JtJ,
-                      Eigen::Matrix<double, 9, 1> &Jtr) const {
-        FactorizedFundamentalMatrix FF(F_cam_pair.F);
-        // Using F directly from ProjectiveImagePair causes issues with U and V flipping signs in some columns
-        const Eigen::Matrix3d F = FF.F();
-
-        // Matrices contain the jacobians of F w.r.t. the factorized fundamental matrix (U,V,sigma)
-        const Eigen::Matrix3d U = quat_to_rotmat(FF.qU);
-        const Eigen::Matrix3d V = quat_to_rotmat(FF.qV);
-
-        Eigen::Matrix3d d_sigma = U.col(1) * V.col(1).transpose();
-        Eigen::Matrix<double, 9, 7> dF_dparams;
-        dF_dparams << 0, F(2, 0), -F(1, 0), 0, F(0, 2), -F(0, 1), d_sigma(0, 0), -F(2, 0), 0, F(0, 0), 0, F(1, 2),
-            -F(1, 1), d_sigma(1, 0), F(1, 0), -F(0, 0), 0, 0, F(2, 2), -F(2, 1), d_sigma(2, 0), 0, F(2, 1), -F(1, 1),
-            -F(0, 2), 0, F(0, 0), d_sigma(0, 1), -F(2, 1), 0, F(0, 1), -F(1, 2), 0, F(1, 0), d_sigma(1, 1), F(1, 1),
-            -F(0, 1), 0, -F(2, 2), 0, F(2, 0), d_sigma(2, 1), 0, F(2, 2), -F(1, 2), F(0, 1), -F(0, 0), 0, d_sigma(0, 2),
-            -F(2, 2), 0, F(0, 2), F(1, 1), -F(1, 0), 0, d_sigma(1, 2), F(1, 2), -F(0, 2), 0, F(2, 1), -F(2, 0), 0,
-            d_sigma(2, 2);
-
-        size_t num_residuals = 0;
-
-        Eigen::Matrix<double, 3, 2> dd1dx1, dd2dx2;
-        dd1dx1 << 1.0, 0.0, 0.0, 1.0, 0.0, 0.0;
-        dd2dx2 << 1.0, 0.0, 0.0, 1.0, 0.0, 0.0;
-
-        double k1 = F_cam_pair.camera1.params[4];
-        double k2 = F_cam_pair.camera2.params[4];
-
-        for (size_t i = 0; i < x1.size(); ++i) {
-            double x1_sq = x1[i].squaredNorm();
-            double x2_sq = x2[i].squaredNorm();
-            Eigen::Matrix<double, 3, 1> d1, d2, h1, h2;
-            d1 << x1[i](0), x1[i](1), 1 + k1 * x1_sq;
-            d2 << x2[i](0), x2[i](1), 1 + k2 * x2_sq;
-            h1 = d1.normalized();
-            h2 = d2.normalized();
-
-            const double C = d2.transpose() * (F * d1);
-
-            dd1dx1(2, 0) = 2 * k1 * x1[i](0);
-            dd1dx1(2, 1) = 2 * k1 * x1[i](1);
-            dd2dx2(2, 0) = 2 * k2 * x2[i](0);
-            dd2dx2(2, 1) = 2 * k2 * x2[i](1);
-
-            Eigen::Matrix<double, 3, 2> J1, J2;
-            J1 = (Eigen::Matrix3d::Identity() - h1 * h1.transpose()) * dd1dx1 / d1.norm();
-            J2 = (Eigen::Matrix3d::Identity() - h2 * h2.transpose()) * dd2dx2 / d2.norm();
-
-            Eigen::Matrix<double, 4, 1> J_C;
-            J_C.block<2, 1>(0, 0) = d2.transpose() * (F * J1);
-            J_C.block<2, 1>(2, 0) = d1.transpose() * (F.transpose() * J2);
-            const double nJ_C = J_C.norm();
-
-            const double inv_nJ_C = 1.0 / nJ_C;
-            const double r = C * inv_nJ_C;
-
-            // Compute weight from robust loss function (used in the IRLS)
-            const double weight = weights[i] * loss_fn.weight(r * r);
-            if (weight == 0.0) {
-                continue;
-            }
-            num_residuals++;
-
-            // Compute Jacobian of Sampson error w.r.t the fundamental/essential matrix (3x3)
-            Eigen::Matrix<double, 1, 9> dF;
-
-            const double s = C * inv_nJ_C * inv_nJ_C;
-
-            dF(0) = (J_C(0) * J1(0, 0) * d2(0));
-            dF(1) = (J_C(0) * J1(0, 0) * d2(1));
-            dF(2) = (J_C(0) * J1(0, 0) * d2(2));
-            dF(3) = (J_C(0) * J1(1, 0) * d2(0));
-            dF(4) = (J_C(0) * J1(1, 0) * d2(1));
-            dF(5) = (J_C(0) * J1(1, 0) * d2(2));
-            dF(6) = (J_C(0) * J1(2, 0) * d2(0));
-            dF(7) = (J_C(0) * J1(2, 0) * d2(1));
-            dF(8) = (J_C(0) * J1(2, 0) * d2(2));
-
-            dF(0) += (J_C(1) * J1(0, 1) * d2(0));
-            dF(1) += (J_C(1) * J1(0, 1) * d2(1));
-            dF(2) += (J_C(1) * J1(0, 1) * d2(2));
-            dF(3) += (J_C(1) * J1(1, 1) * d2(0));
-            dF(4) += (J_C(1) * J1(1, 1) * d2(1));
-            dF(5) += (J_C(1) * J1(1, 1) * d2(2));
-            dF(6) += (J_C(1) * J1(2, 1) * d2(0));
-            dF(7) += (J_C(1) * J1(2, 1) * d2(1));
-            dF(8) += (J_C(1) * J1(2, 1) * d2(2));
-
-            dF(0) += (J_C(2) * J2(0, 0) * d1(0));
-            dF(1) += (J_C(2) * J2(1, 0) * d1(0));
-            dF(2) += (J_C(2) * J2(2, 0) * d1(0));
-            dF(3) += (J_C(2) * J2(0, 0) * d1(1));
-            dF(4) += (J_C(2) * J2(1, 0) * d1(1));
-            dF(5) += (J_C(2) * J2(2, 0) * d1(1));
-            dF(6) += (J_C(2) * J2(0, 0) * d1(2));
-            dF(7) += (J_C(2) * J2(1, 0) * d1(2));
-            dF(8) += (J_C(2) * J2(2, 0) * d1(2));
-
-            dF(0) += (J_C(3) * J2(0, 1) * d1(0));
-            dF(1) += (J_C(3) * J2(1, 1) * d1(0));
-            dF(2) += (J_C(3) * J2(2, 1) * d1(0));
-            dF(3) += (J_C(3) * J2(0, 1) * d1(1));
-            dF(4) += (J_C(3) * J2(1, 1) * d1(1));
-            dF(5) += (J_C(3) * J2(2, 1) * d1(1));
-            dF(6) += (J_C(3) * J2(0, 1) * d1(2));
-            dF(7) += (J_C(3) * J2(1, 1) * d1(2));
-            dF(8) += (J_C(3) * J2(2, 1) * d1(2));
-
-            dF *= -s;
-
-            dF(0) += d1(0) * d2(0);
-            dF(1) += d1(0) * d2(1);
-            dF(2) += d1(0) * d2(2);
-            dF(3) += d1(1) * d2(0);
-            dF(4) += d1(1) * d2(1);
-            dF(5) += d1(1) * d2(2);
-            dF(6) += d1(2) * d2(0);
-            dF(7) += d1(2) * d2(1);
-            dF(8) += d1(2) * d2(2);
-
-            dF *= inv_nJ_C;
-
-            // dLdk = dCdk * den / (den ** 2) + next
-            double dLdk1 = inv_nJ_C * F.col(2).dot(d2) * x1_sq;
-            double dLdk2 = inv_nJ_C * F.row(2).dot(d1) * x2_sq;
-
-            double s1 = d1(2) * d1(2) + x1_sq;
-            double s2 = d2(2) * d2(2) + x2_sq;
-            Eigen::Matrix<double, 3, 2> dJ1dk, dJ2dk;
-            double d10_sq = d1(0) * d1(0), d11_sq = d1(1) * d1(1);
-            dJ1dk(0, 0) = (3 * d1(2) * x1_sq * (2 * d10_sq * d1(2) * k1 + d10_sq - s1) -
-                           2 * s1 * (d10_sq * d1(2) + d10_sq * k1 * x1_sq - d1(2) * x1_sq));
-            dJ1dk(0, 1) = d1(0) * d1(1) * (3 * d1(2) * x1_sq * (2 * d1(2) * k1 + 1) - 2 * s1 * (2 * k1 * x1_sq + 1));
-            dJ1dk(1, 0) = dJ1dk(0, 1);
-            dJ1dk(1, 1) = (3 * d1(2) * x1_sq * (2 * d11_sq * d1(2) * k1 + d11_sq - s1) -
-                           2 * s1 * (d11_sq * d1(2) + d11_sq * k1 * x1_sq - d1(2) * x1_sq));
-            dJ1dk(2, 0) = d1(0);
-            dJ1dk(2, 1) = d1(1);
-            dJ1dk.row(2) *= -x1_sq * (3 * d1(2) * (d1(2) - 2) - s1);
-            dJ1dk /= std::pow(s1, 2.5);
-
-            double d20_sq = d2(0) * d2(0), d21_sq = d2(1) * d2(1);
-            dJ2dk(0, 0) = (3 * d2(2) * x2_sq * (2 * d20_sq * d2(2) * k2 + d20_sq - s2) -
-                           2 * s2 * (d20_sq * d2(2) + d20_sq * k2 * x2_sq - d2(2) * x2_sq));
-            dJ2dk(0, 1) = d2(0) * d2(1) * (3 * d2(2) * x2_sq * (2 * d2(2) * k2 + 1) - 2 * s2 * (2 * k2 * x2_sq + 1));
-            dJ2dk(1, 0) = dJ2dk(0, 1);
-            dJ2dk(1, 1) = (3 * d2(2) * x2_sq * (2 * d21_sq * d2(2) * k2 + d21_sq - s2) -
-                           2 * s2 * (d21_sq * d2(2) + d21_sq * k2 * x2_sq - d2(2) * x2_sq));
-            dJ2dk(2, 0) = d2(0);
-            dJ2dk(2, 1) = d2(1);
-            dJ2dk.row(2) *= -x2_sq * (3 * d2(2) * (d2(2) - 2) - s2);
-            dJ2dk /= std::pow(s2, 2.5);
-
-            Eigen::Matrix<double, 3, 1> dd1dk, dd2dk;
-            dd1dk << 0.0, 0.0, x1_sq;
-            dd2dk << 0.0, 0.0, x2_sq;
-
-            dLdk1 -= C * inv_nJ_C * inv_nJ_C * inv_nJ_C *
-                     (d2.transpose() * F * J1 * (d2.transpose() * F * dJ1dk).transpose() +
-                      d1.transpose() * F.transpose() * J2 * (x1_sq * F.col(2).transpose() * J2).transpose())(0, 0);
-
-            dLdk2 -= C * inv_nJ_C * inv_nJ_C * inv_nJ_C *
-                     (d2.transpose() * F * J1 * (x2_sq * F.row(2) * J1).transpose() +
-                      d1.transpose() * F.transpose() * J2 * (d1.transpose() * F.transpose() * dJ2dk).transpose())(0, 0);
-
-            // and then w.r.t. the pose parameters (rotation + tangent basis for translation + k)
-            Eigen::Matrix<double, 1, 9> J;
-            J.block<1, 7>(0, 0) = dF * dF_dparams;
-            J(0, 7) = dLdk1;
-            J(0, 8) = dLdk2;
-
-            // Accumulate into JtJ and Jtr
-            Jtr += weight * C * inv_nJ_C * J.transpose();
-            for (size_t row = 0; row < 9; ++row) {
-                for (size_t col = 0; col <= row; ++col) {
-                    JtJ(row, col) += weight * (J(row) * J(col));
-                }
-            }
-        }
-        return num_residuals;
-    }
-
-    ProjectiveImagePair step(Eigen::Matrix<double, 9, 1> dp, const ProjectiveImagePair &F_cam_pair) const {
-        FactorizedFundamentalMatrix F = FactorizedFundamentalMatrix(F_cam_pair.F);
-        FactorizedFundamentalMatrix F_new;
-
-        F_new.qU = quat_step_pre(F.qU, dp.block<3, 1>(0, 0));
-        F_new.qV = quat_step_pre(F.qV, dp.block<3, 1>(3, 0));
-        F_new.sigma = F.sigma + dp(6);
-
-        Camera camera1_new = Camera(
-            "DIVISION", std::vector<double>{1.0, 1.0, 0.0, 0.0, F_cam_pair.camera1.params[4] + dp(7, 0)}, -1, -1);
-        Camera camera2_new = Camera(
-            "DIVISION", std::vector<double>{1.0, 1.0, 0.0, 0.0, F_cam_pair.camera2.params[4] + dp(8, 0)}, -1, -1);
-
-        return ProjectiveImagePair(F_new.F(), camera1_new, camera2_new);
-    }
-
-    typedef ProjectiveImagePair param_t;
-    static constexpr size_t num_params = 9;
-
-  private:
-    const std::vector<Point2D> &x1;
-    const std::vector<Point2D> &x2;
-    const LossFunction &loss_fn;
-    const ResidualWeightVector &weights;
-};
-
-// Non-linear refinement of transfer error |x2 - pi(H*x1)|^2, parameterized by fixing H(2,2) = 1
-// I did some preliminary experiments comparing different error functions (e.g. symmetric and transfer)
-// as well as other parameterizations (different affine patches, SVD as in Bartoli/Sturm, etc)
-// but it does not seem to have a big impact (and is sometimes even worse)
-// Implementations of these can be found at https://github.com/vlarsson/homopt
-template <typename LossFunction, typename ResidualWeightVector = UniformWeightVector>
-class HomographyJacobianAccumulator {
-  public:
-    HomographyJacobianAccumulator(const std::vector<Point2D> &points2D_1, const std::vector<Point2D> &points2D_2,
-                                  const LossFunction &l, const ResidualWeightVector &w = ResidualWeightVector())
-        : x1(points2D_1), x2(points2D_2), loss_fn(l), weights(w) {}
-
-    double residual(const Eigen::Matrix3d &H) const {
-        double cost = 0.0;
-
-        const double H0_0 = H(0, 0), H0_1 = H(0, 1), H0_2 = H(0, 2);
-        const double H1_0 = H(1, 0), H1_1 = H(1, 1), H1_2 = H(1, 2);
-        const double H2_0 = H(2, 0), H2_1 = H(2, 1), H2_2 = H(2, 2);
-
-        for (size_t k = 0; k < x1.size(); ++k) {
-            const double x1_0 = x1[k](0), x1_1 = x1[k](1);
-            const double x2_0 = x2[k](0), x2_1 = x2[k](1);
-
-            const double Hx1_0 = H0_0 * x1_0 + H0_1 * x1_1 + H0_2;
-            const double Hx1_1 = H1_0 * x1_0 + H1_1 * x1_1 + H1_2;
-            const double inv_Hx1_2 = 1.0 / (H2_0 * x1_0 + H2_1 * x1_1 + H2_2);
-
-            const double r0 = Hx1_0 * inv_Hx1_2 - x2_0;
-            const double r1 = Hx1_1 * inv_Hx1_2 - x2_1;
-            const double r2 = r0 * r0 + r1 * r1;
-            cost += weights[k] * loss_fn.loss(r2);
-        }
-        return cost;
-    }
-
-    size_t accumulate(const Eigen::Matrix3d &H, Eigen::Matrix<double, 8, 8> &JtJ, Eigen::Matrix<double, 8, 1> &Jtr) {
-        Eigen::Matrix<double, 2, 8> dH;
-        const double H0_0 = H(0, 0), H0_1 = H(0, 1), H0_2 = H(0, 2);
-        const double H1_0 = H(1, 0), H1_1 = H(1, 1), H1_2 = H(1, 2);
-        const double H2_0 = H(2, 0), H2_1 = H(2, 1), H2_2 = H(2, 2);
-
-        size_t num_residuals = 0;
-        for (size_t k = 0; k < x1.size(); ++k) {
-            const double x1_0 = x1[k](0), x1_1 = x1[k](1);
-            const double x2_0 = x2[k](0), x2_1 = x2[k](1);
-
-            const double Hx1_0 = H0_0 * x1_0 + H0_1 * x1_1 + H0_2;
-            const double Hx1_1 = H1_0 * x1_0 + H1_1 * x1_1 + H1_2;
-            const double inv_Hx1_2 = 1.0 / (H2_0 * x1_0 + H2_1 * x1_1 + H2_2);
-
-            const double z0 = Hx1_0 * inv_Hx1_2;
-            const double z1 = Hx1_1 * inv_Hx1_2;
-
-            const double r0 = z0 - x2_0;
-            const double r1 = z1 - x2_1;
-            const double r2 = r0 * r0 + r1 * r1;
-
-            // Compute weight from robust loss function (used in the IRLS)
-            const double weight = weights[k] * loss_fn.weight(r2);
-            if (weight == 0.0)
-                continue;
-            num_residuals++;
-
-            dH << x1_0, 0.0, -x1_0 * z0, x1_1, 0.0, -x1_1 * z0, 1.0, 0.0, // -z0,
-                0.0, x1_0, -x1_0 * z1, 0.0, x1_1, -x1_1 * z1, 0.0, 1.0;   // -z1,
-            dH = dH * inv_Hx1_2;
-
-            // accumulate into JtJ and Jtr
-            Jtr += dH.transpose() * (weight * Eigen::Vector2d(r0, r1));
-            for (size_t i = 0; i < 8; ++i) {
-                for (size_t j = 0; j <= i; ++j) {
-                    JtJ(i, j) += weight * dH.col(i).dot(dH.col(j));
-                }
-            }
-        }
-        return num_residuals;
-    }
-
-    Eigen::Matrix3d step(Eigen::Matrix<double, 8, 1> dp, const Eigen::Matrix3d &H) const {
-        Eigen::Matrix3d H_new = H;
-        Eigen::Map<Eigen::Matrix<double, 8, 1>>(H_new.data()) += dp;
-        return H_new;
-    }
-    typedef Eigen::Matrix3d param_t;
-    static constexpr size_t num_params = 8;
-
-  private:
-    const std::vector<Point2D> &x1;
-    const std::vector<Point2D> &x2;
-    const LossFunction &loss_fn;
-    const ResidualWeightVector &weights;
-};
-
-template <typename LossFunction, typename ResidualWeightVector = UniformWeightVector>
-class Radial1DJacobianAccumulator {
-  public:
-    Radial1DJacobianAccumulator(const std::vector<Point2D> &points2D, const std::vector<Point3D> &points3D,
-                                const LossFunction &l, const ResidualWeightVector &w = ResidualWeightVector())
-        : x(points2D), X(points3D), loss_fn(l), weights(w) {}
-
-    double residual(const CameraPose &pose) const {
-        double cost = 0.0;
-        Eigen::Matrix3d R = pose.R();
-        for (size_t k = 0; k < x.size(); ++k) {
-            Eigen::Vector2d z = (R * X[k] + pose.t).template topRows<2>().normalized();
-            double alpha = z.dot(x[k]);
-            // This assumes points will not cross the half-space during optimization
-            if (alpha < 0)
-                continue;
-            double r2 = (alpha * z - x[k]).squaredNorm();
-            cost += weights[k] * loss_fn.loss(r2);
-        }
-
-        return cost;
-    }
-
-    size_t accumulate(const CameraPose &pose, Eigen::Matrix<double, 5, 5> &JtJ,
-                      Eigen::Matrix<double, 5, 1> &Jtr) const {
-        Eigen::Matrix3d R = pose.R();
-        size_t num_residuals = 0;
-        for (size_t k = 0; k < x.size(); ++k) {
-            Eigen::Vector3d RX = R * X[k];
-            const Eigen::Vector2d z = (RX + pose.t).topRows<2>();
-
-            const double n_z = z.norm();
-            const Eigen::Vector2d zh = z / n_z;
-            const double alpha = zh.dot(x[k]);
-            // This assumes points will not cross the half-space during optimization
-            if (alpha < 0)
-                continue;
-
-            // Setup residual
-            Eigen::Vector2d r = alpha * zh - x[k];
-            const double r_squared = r.squaredNorm();
-            const double weight = weights[k] * loss_fn.weight(r_squared);
-
-            if (weight == 0.0) {
-                continue;
-            }
-            num_residuals++;
-
-            // differentiate residual with respect to z
-            Eigen::Matrix2d dr_dz = (zh * x[k].transpose() + alpha * Eigen::Matrix2d::Identity()) *
-                                    (Eigen::Matrix2d::Identity() - zh * zh.transpose()) / n_z;
-
-            Eigen::Matrix<double, 2, 5> dz;
-            dz << 0.0, RX(2), -RX(1), 1.0, 0.0, -RX(2), 0.0, RX(0), 0.0, 1.0;
-
-            Eigen::Matrix<double, 2, 5> J = dr_dz * dz;
-
-            // Accumulate into JtJ and Jtr
-            Jtr += weight * J.transpose() * r;
-            for (size_t i = 0; i < 5; ++i) {
-                for (size_t j = 0; j <= i; ++j) {
-                    JtJ(i, j) += weight * (J.col(i).dot(J.col(j)));
-                }
-            }
-        }
-        return num_residuals;
-    }
-
-    CameraPose step(Eigen::Matrix<double, 5, 1> dp, const CameraPose &pose) const {
-        CameraPose pose_new;
-        pose_new.q = quat_step_pre(pose.q, dp.block<3, 1>(0, 0));
-        pose_new.t(0) = pose.t(0) + dp(3);
-        pose_new.t(1) = pose.t(1) + dp(4);
-        return pose_new;
-    }
-    typedef CameraPose param_t;
-    static constexpr size_t num_params = 5;
-
-  private:
-    const std::vector<Point2D> &x;
-    const std::vector<Point3D> &X;
-    const LossFunction &loss_fn;
-    const ResidualWeightVector &weights;
-};
-
-} // namespace poselib
-
-#endif
